file_server/requirements.txt
aiofiles
grpcio
grpcio-tools
pytest